+++ PlanBuild/Blueprints/BlueprintManager.cs
@@ -430,11 +430,12 @@ namespace PlanBuild.Blueprints
         /// <param name="go"></param>
         private void UITooltip_OnHoverStart(On.UITooltip.orig_OnHoverStart orig, UITooltip self, GameObject go)
         {
-            if (BlueprintAssets.BlueprintTooltip && BlueprintConfig.TooltipEnabledConfig.Value &&
-                Hud.instance.m_hoveredPiece && Hud.instance.m_hoveredPiece.name.StartsWith(Blueprint.PieceBlueprintName))
+            if (BlueprintAssets.BlueprintTooltip && Hud.instance.m_hoveredPiece is Piece piece)
             {
-                string bpname = Hud.instance.m_hoveredPiece.name.Substring(Blueprint.PieceBlueprintName.Length + 1);
-                if (LocalBlueprints.TryGetValue(bpname, out var bp) && bp.Thumbnail != null)
+                if (BlueprintConfig.TooltipEnabledConfig.Value &&
+                    Hud.instance.m_hoveredPiece.name.StartsWith(Blueprint.PieceBlueprintName) &&
+                    LocalBlueprints.TryGetValue(piece.name.Substring(Blueprint.PieceBlueprintName.Length + 1), out var bp) && 
+                    bp.Thumbnail != null)
                 {
                     self.m_tooltipPrefab = BlueprintAssets.BlueprintTooltip;
                     orig(self, go);
@@ -442,11 +443,16 @@ namespace PlanBuild.Blueprints
                         .GetComponent<Image>().color = BlueprintConfig.TooltipBackgroundConfig.Value;
                     global::Utils.FindChild(UITooltip.m_tooltip.transform, "Image")
                         .GetComponent<Image>().sprite = Sprite.Create(bp.Thumbnail, new Rect(0, 0, bp.Thumbnail.width, bp.Thumbnail.height), Vector2.zero);
-                    return;
                 }
-            }
+                else
+                {
+                    self.m_tooltipPrefab = OriginalTooltip;
+                    orig(self, go);
+                }
 
-            self.m_tooltipPrefab = OriginalTooltip;
+                return;
+            }
+            
             orig(self, go);
         }
     }